+++ spec/lite/ruby/configuration_spec.rb
@@ -5,15 +5,17 @@ require 'spec_helper'
 RSpec.describe Lite::Ruby::Configuration do
   after do
     Lite::Ruby.configure do |config|
-      config.array = true
+      config.monkey_patches = %w[
+        array date enumerable hash integer kernel numeric object range string struct time
+      ]
     end
   end
 
   describe '#configure' do
     it 'to be "foo"' do
-      Lite::Ruby.configuration.array = 'foo'
+      Lite::Ruby.configuration.monkey_patches = 'foo'
 
-      expect(Lite::Ruby.configuration.array).to eq('foo')
+      expect(Lite::Ruby.configuration.monkey_patches).to eq('foo')
     end
   end