+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [1.0.1] - 2019-07-12
+# Changed
+- Rename `multiply => product` and `divisible => quotient`
+
+## [1.0.1] - 2019-07-12
 ### Added
 - Added String => `acronym`
 - Added Struct extensions